requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Pillow==9.4.0
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.1
reportlab==3.6.12
scrypt==0.8.20
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
